Added consts file to host constants for scrapers
